refactor(uffd): extract Prefault to its own file
Move Prefault and the directDataSource helper from userfaultfd.go
into a dedicated prefault.go. Pure file move, no behavior change.
